+++ lib/src/async/device.dart
@@ -66,7 +66,6 @@ class Device {
       _handler.device.buildPullFolderRequest(path),
       _handler.device.parsePullFolderResponse);
 
-
   /// Get clipboard
   ///
   /// Returns base64 encoded string. You must decode it properly.
@@ -76,12 +75,15 @@ class Device {
   ///     // Returns a string which is base64 encoded
   ///     utf8.decode(base64.decode(await driver.device.getClipboard()))
   ///
-  Future<String> getClipboard({String contentType = ContentType.plaintext}) => _client.send(
-      _handler.device.buildGetClipboardRequest(contentType: contentType),
-      _handler.device.parseGetClipboardResponse);
+  Future<String> getClipboard({String contentType = ContentType.plaintext}) =>
+      _client.send(
+          _handler.device.buildGetClipboardRequest(contentType: contentType),
+          _handler.device.parseGetClipboardResponse);
 
   /// Set clipboard
   /// Sends base64 encoded string. You must decode it properly as the 1st argument.
+  ///
+  /// `label` argument is only for Android like `'Note'`
   ///
   /// For example:
   ///
@@ -89,9 +91,12 @@ class Device {
   ///     await driver.device.setClipboard(
   ///         base64.encode(utf8.encode('happy testing')));
   ///
-  Future<void> setClipboard(String base64encoded, {String contentType = ContentType.plaintext}) => _client.send(
-      _handler.device.buildSetClipboardRequest(base64encoded, contentType: contentType),
-      _handler.device.parseSetClipboardResponse);
+  Future<void> setClipboard(String base64encoded,
+          {String contentType = ContentType.plaintext, String label}) =>
+      _client.send(
+          _handler.device.buildSetClipboardRequest(base64encoded,
+              contentType: contentType, label: label),
+          _handler.device.parseSetClipboardResponse);
 
   @override
   int get hashCode => _client.hashCode;