+++ lib/src/common/webdriver_handler.dart
@@ -1,5 +1,4 @@
 import 'dart:math';
-import 'dart:convert';
 
 import 'package:appium_driver/src/common/app_state.dart';
 import 'package:appium_driver/src/common/clipboard.dart';
@@ -736,15 +735,15 @@ abstract class DeviceHandler {
   String parsePullFolderResponse(WebDriverResponse response);
 
   /// Builds request for 'get clipboard'
-  WebDriverRequest buildGetClipboardRequest({String contentType = ContentType.plaintext});
+  WebDriverRequest buildGetClipboardRequest(
+      {String contentType = ContentType.plaintext});
 
   /// Parse response for 'get clipboard'
   String parseGetClipboardResponse(WebDriverResponse response);
 
-
   /// Builds request for 'set clipboard'
   WebDriverRequest buildSetClipboardRequest(String base64encoded,
-      {String contentType = ContentType.plaintext});
+      {String contentType = ContentType.plaintext, String label});
 
   /// Parse response for 'set clipboard'
   void parseSetClipboardResponse(WebDriverResponse response);