+++ test/functional/appium_dart_test.dart
@@ -83,16 +83,14 @@ void main() {
 
     await driver.device.pushFile('AddressBook.png', pulledFile);
   });
-  
-  test('clipboard', () async {
 
-    await driver.device.setClipboard(
-        base64.encode(utf8.encode('happy testing')));
+  test('clipboard', () async {
+    await driver.device
+        .setClipboard(base64.encode(utf8.encode('happy testing')));
     expect(utf8.decode(base64.decode(await driver.device.getClipboard())),
         'happy testing');
 
-    await driver.device.setClipboard(
-        base64.encode(utf8.encode('appium')));
+    await driver.device.setClipboard(base64.encode(utf8.encode('appium')));
     expect(utf8.decode(base64.decode(await driver.device.getClipboard())),
         'appium');
   });